Modificacion en el service

+++ Service/ComboService.cs
@@ -2,6 +2,7 @@
 using LiamellCruz_Ap1_P2.Models;
 using Microsoft.EntityFrameworkCore;
 using System.Linq.Expressions;
+using static Azure.Core.HttpHeader;
 
 namespace LiamellCruz_Ap1_P2.Service;
 
@@ -9,7 +10,7 @@ public class ComboService(IDbContextFactory<Contexto> DbFactory)
 {
     public async Task<bool> Guardar(Combo combo)
     {
-        await using var contexto = await DbFactory.CreateDbContextAsync();
+        
         if (!await Existe(combo.ComboId))
             return await Insertar(combo);
         else
@@ -23,12 +24,15 @@ public class ComboService(IDbContextFactory<Contexto> DbFactory)
             .AnyAsync(c => c.ComboId == id);
     }
 
+
     private async Task<bool> Insertar(Combo combo)
     {
         await using var contexto = await DbFactory.CreateDbContextAsync();
+        await AfectarCantidad(combo.ComboDetalle.ToArray(), true);
         contexto.Combo.Add(combo);
-        return await contexto.SaveChangesAsync() > 0;
+        return await contexto.SaveChangesAsync() > 0; ;
     }
+
 
     private async Task<bool> Modificar(Combo combo)
     {
@@ -39,14 +43,34 @@ public class ComboService(IDbContextFactory<Contexto> DbFactory)
         return modificado;
     }
 
-   
+
     public async Task<bool> Eliminar(int id)
     {
         await using var contexto = await DbFactory.CreateDbContextAsync();
+        var combo = contexto.Combo.Find(id);
+        if (combo == null)
+            return false;
+
+
+        await AfectarCantidad(combo.ComboDetalle.ToArray(), false);
         return await contexto.Combo
-           .AsNoTracking()
-           .Where(c => c.ComboId == id)
-           .ExecuteDeleteAsync() > 0;
+            .Include(c => c.ComboDetalle)
+            .Where(c => c.ComboId == combo.ComboId)
+            .ExecuteDeleteAsync() > 0;
+    }
+
+    public async Task AfectarCantidad(ComboDetalle[] detalles, bool resta)
+    {
+        await using var contexto = await DbFactory.CreateDbContextAsync();
+        foreach (var item in detalles)
+        {
+            var detalle = await contexto.ComboDetalle.SingleAsync(d => d.DetalleId == item.DetalleId);
+            if (resta)
+                detalle.Cantidad -= item.Cantidad;
+            else
+                detalle.Cantidad += item.Cantidad;
+        }
+        await contexto.SaveChangesAsync();
     }
 
     public async Task<Combo> Buscar(int id)
@@ -62,16 +86,12 @@ public class ComboService(IDbContextFactory<Contexto> DbFactory)
     {
         await using var contexto = await DbFactory.CreateDbContextAsync();
         return await contexto.Combo
-            .AsNoTracking()
-            .Where(criterio)
-            .ToListAsync();
-
-
-
-
+        .Include(c => c.ComboDetalle)
+        .AsNoTracking()
+        .Where(criterio)
+        .ToListAsync();
     }
+
 }
-
-
